Fallback to empty string when metadata name is invalid
