+++ packages/twenty-front/src/modules/object-metadata/validation-schemas/metadataLabelSchema.ts
@@ -1,8 +1,8 @@
 import { errors } from '@/settings/data-model/fields/forms/utils/errorMessages';
 import { z } from 'zod';
+import { computeMetadataNameFromLabelOrThrow } from '~/pages/settings/data-model/utils/compute-metadata-name-from-label-or-throw.utils';
 
 import { METADATA_LABEL_VALID_PATTERN } from '~/pages/settings/data-model/constants/MetadataLabelValidPattern';
-import { computeMetadataNameFromLabel } from '~/pages/settings/data-model/utils/compute-metadata-name-from-label.utils';
 export const metadataLabelSchema = (existingLabels?: string[]) => {
   return z
     .string()
@@ -12,7 +12,7 @@ export const metadataLabelSchema = (existingLabels?: string[]) => {
     .refine(
       (label) => {
         try {
-          computeMetadataNameFromLabel(label);
+          computeMetadataNameFromLabelOrThrow(label);
           return true;
         } catch {
           return false;
@@ -28,7 +28,9 @@ export const metadataLabelSchema = (existingLabels?: string[]) => {
           if (!existingLabels || !label?.length) {
             return true;
           }
-          return !existingLabels.includes(computeMetadataNameFromLabel(label));
+          return !existingLabels.includes(
+            computeMetadataNameFromLabelOrThrow(label),
+          );
         } catch {
           return false;
         }
