+++ packages/twenty-front/src/pages/settings/data-model/utils/compute-metadata-name-from-label-or-throw.utils.ts
@@ -1,0 +1,22 @@
+import camelCase from 'lodash.camelcase';
+import { slugify } from 'transliteration';
+
+export const computeMetadataNameFromLabelOrThrow = (label: string): string => {
+  const prefixedLabel = /^\d/.test(label) ? `n${label}` : label;
+
+  if (prefixedLabel === '') {
+    return '';
+  }
+
+  const formattedString = slugify(prefixedLabel, {
+    trim: true,
+    separator: '_',
+    allowedChars: 'a-zA-Z0-9',
+  });
+
+  if (formattedString === '') {
+    throw new Error('Invalid label');
+  }
+
+  return camelCase(formattedString);
+};
